move extractors/viv/__init__.py logic to extractor.py
