+++ capa/features/extractors/viv/extractor.py
@@ -1,0 +1,77 @@
+# Copyright (C) 2020 FireEye, Inc. All Rights Reserved.
+# Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at: [package root]/LICENSE.txt
+# Unless required by applicable law or agreed to in writing, software distributed under the License
+#  is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and limitations under the License.
+import logging
+
+import viv_utils
+import viv_utils.flirt
+
+import capa.features.extractors
+import capa.features.extractors.viv.file
+import capa.features.extractors.viv.insn
+import capa.features.extractors.viv.function
+import capa.features.extractors.viv.basicblock
+from capa.features.extractors import FeatureExtractor
+
+logger = logging.getLogger(__name__)
+
+
+class InstructionHandle:
+    """this acts like a vivisect.Opcode but with an __int__() method"""
+
+    def __init__(self, inner):
+        self._inner = inner
+
+    def __int__(self):
+        return self.va
+
+    def __getattr__(self, name):
+        return getattr(self._inner, name)
+
+
+class VivisectFeatureExtractor(FeatureExtractor):
+    def __init__(self, vw, path):
+        super(VivisectFeatureExtractor, self).__init__()
+        self.vw = vw
+        self.path = path
+
+    def get_base_address(self):
+        # assume there is only one file loaded into the vw
+        return list(self.vw.filemeta.values())[0]["imagebase"]
+
+    def extract_file_features(self):
+        for feature, va in capa.features.extractors.viv.file.extract_features(self.vw, self.path):
+            yield feature, va
+
+    def get_functions(self):
+        for va in sorted(self.vw.getFunctions()):
+            yield viv_utils.Function(self.vw, va)
+
+    def extract_function_features(self, f):
+        for feature, va in capa.features.extractors.viv.function.extract_features(f):
+            yield feature, va
+
+    def get_basic_blocks(self, f):
+        return f.basic_blocks
+
+    def extract_basic_block_features(self, f, bb):
+        for feature, va in capa.features.extractors.viv.basicblock.extract_features(f, bb):
+            yield feature, va
+
+    def get_instructions(self, f, bb):
+        for insn in bb.instructions:
+            yield InstructionHandle(insn)
+
+    def extract_insn_features(self, f, bb, insn):
+        for feature, va in capa.features.extractors.viv.insn.extract_features(f, bb, insn):
+            yield feature, va
+
+    def is_library_function(self, va):
+        return viv_utils.flirt.is_library_function(self.vw, va)
+
+    def get_function_name(self, va):
+        return viv_utils.get_function_name(self.vw, va)
